refactoring(LottoResult): getYield 가 테스트와 구현이 모두 이상한 것을 발견하고 수정
getYield 수정 :
  - getYield TC에 사용한 값을 나눈 값으로 수정
  - LottoResult 에 spentMoney값을 입력하도록 수정
  - getYield에 spentMoney로 나눈 값으로 수익률을 받도록 수정
  -

+++ src/main/java/lotto/domain/LottoResult.java
@@ -6,15 +6,15 @@ import java.util.Objects;
 
 public class LottoResult {
 
-	private final Map<LottoRank, Integer> winningMap;
+	private final Map<LottoRank, Long> winningMap;
 	private final int buyPrize;
 
-	public LottoResult(Map<LottoRank, Integer> winningMap, int buyPrize) {
+	public LottoResult(Map<LottoRank, Long> winningMap, int buyPrize) {
 		this.winningMap = Collections.unmodifiableMap(winningMap);
 		this.buyPrize = buyPrize;
 	}
 
-	public Map<LottoRank, Integer> getWinningMap() {
+	public Map<LottoRank, Long> getWinningMap() {
 		return winningMap;
 	}
 
@@ -35,7 +35,7 @@ public class LottoResult {
 		return winningMap
 				.entrySet()
 				.stream()
-				.mapToInt(entry -> entry.getKey().getEarningPrize() * entry.getValue())
-				.sum()/(double)buyPrize;
+				.mapToLong(entry -> entry.getKey().getEarningPrize() * entry.getValue())
+				.sum() / (double) buyPrize;
 	}
 }
